extract chachapoly logic in cifra so that it can be reused
